refactor: relocate product API and adapter files to adapters/product/ directory
