+++ adapters/product/product-adapter.ts
